requirements.txt
flask==3.1.1
gunicorn==23.0.0
flask-sqlalchemy
Pillow
google-generativeai
